pyproject.toml
[tool.poetry]
name = "tabby"
version = "0.1.0"
description = "A tiny Discord bot for servers that I like"
authors = ["Kaylynn <mkaylynn7@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.2.2"
asyncpg = "^0.27.0"
toml = "^0.10.2"
asyncpg-stubs = "^0.27.0"
aiohttp = "^3.8.4"
typing-extensions = "^4.5.0"
selenium = "^4.8.3"
yarl = "^1.8.2"
prettytable = "^3.6.0"
python-slugify = "^8.0.1"
pydantic = "^1.10.7"
cryptography = "^40.0.2"
jinja2 = "^3.1.2"
mistletoe = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
